+++ CHANGELOG.md
@@ -2,5 +2,13 @@
 
 All notable changes to `permissions-admin` will be documented in this file.
 
+## 1.0.2 - 2024-05-08
+- Fixes the UserComponent modal and the form.
+- Added the option to disable eding the User Name, especially if are using the User model as first_name, last_name instead of name.
+
+## 1.0.1 - 2024-05-7
+- Fix the scroll in the modal when the data is out of the screen.
+
 ## 1.0.0 - 2024-05-07
 - Initial release.
+